Phase 7b audit S3: keep intensity expectation in fp32 under bf16 autocast
probs * centers where centers[-1]=50 has ~0.4 mm quantization noise in
bf16 — comparable to the CSI@30mm decision boundary error budget. Cast
softmax output and band_centers to fp32 before the per-band weighted sum
so the expected-value rain map lives in fp32 regardless of autocast dtype.
F.cross_entropy upcasts internally, so the logits-side loss is unchanged.
Test verifies (a) the expectation tensor is fp32 under bf16 autocast and
(b) the fp32-multiply path strictly improves on the all-bf16 baseline.

+++ model/decoder.py
@@ -219,8 +219,14 @@ class PluvianDecoder(nn.Module):
         if self.intensity_stratified:
             logits = rearrange(rain_out, "(b t) k h w -> b t k h w",
                                b=B, t=self.forecast_frames)
-            probs = F.softmax(logits, dim=2)
-            centers = self.band_centers.view(1, 1, -1, 1, 1)
+            # Phase 7b audit S3: bf16 softmax + (probs*centers) where the top
+            # band centre is 50 mm/h gives ~0.4 mm quantization noise — that
+            # is the same order as the CSI@30mm decision boundary. Cast to
+            # fp32 for the expectation so the head topology is not what we
+            # are measuring at the high-intensity end. F.cross_entropy already
+            # upcasts internally, so logits stay in their autocast dtype.
+            probs = F.softmax(logits, dim=2).float()
+            centers = self.band_centers.float().view(1, 1, -1, 1, 1)
             rain = (probs * centers).sum(dim=2)
             return rain, pwv, logits
         rain = rearrange(rain_out, "(b t) () h w -> b t h w",
